bring `ref_info` Commit into the right position
